+++ src/SIL.Machine.WebApi/Models/WordGraphDto.cs
@@ -4,7 +4,7 @@ namespace SIL.Machine.WebApi.Models
 {
 	public class WordGraphDto
 	{
-		public double InitialStateScore { get; set; }
+		public float InitialStateScore { get; set; }
 		public IReadOnlyList<int> FinalStates { get; set; }
 		public IReadOnlyList<WordGraphArcDto> Arcs { get; set; }
 	}
